Refactor entire app.component.html to use separate EmployeesComponent and FooterComponent - EmployeesComponent will now contain all the logic for showing - Fixed "NullInjectorError: No provider for {token} found!" error (https://angular.io/errors/NG0201) - Error appeared (I guess) because the header.component.ts is using EmployeesComponent in its contructor in order the onOpenModal() and searchEmployees() methods
