+++ employeemanager-angular/src/app/app.component.html
@@ -3,40 +3,8 @@
 
 <!-- Container -->
 <div class="container" id="main-container">
-  <div class="row">
-    <!-- Employee card -->
-    <app-employee-card class="col-md-6 col-xl-3" *ngFor="let employee of employees" [employee]="employee">
-    </app-employee-card>
-  </div>
-</div>
-
-<!-- Add Employee Modal -->
-<app-add-employee-modal class="modal fade" id="addEmployeeModal" tabindex="-1" role="dialog"
-  aria-labelledby="addEmployeeModalLabel" aria-hidden="true"></app-add-employee-modal>
-
-<!-- Edit Modal -->
-<app-edit-employee-modal [editEmployee]="editEmployee" class="modal fade" id="editEmployeeModal" tabindex="-1"
-  role="dialog" aria-labelledby="employeeEditModalLabel" aria-hidden="true"></app-edit-employee-modal>
-
-<!-- Delete Modal -->
-<app-delete-employee-modal [deleteEmployee]="deleteEmployee" class="modal fade" id="deleteEmployeeModal" tabindex="-1"
-  role="dialog" aria-labelledby="deleteModelLabel" aria-hidden="true"></app-delete-employee-modal>
-
-<!-- Notification for no employees -->
-<div *ngIf="employees?.length === 0" class="col-lg-12 col-md-12 col-xl-12">
-  <div class="alert alert-info" role="alert">
-    <div class="alert-heading alert-title">NO EMPLOYEES!</div>
-    <p>No Employees were found.</p>
-  </div>
+  <app-employees></app-employees>
 </div>
 
 <!-- Footer -->
-<footer class="page-footer font-small blue">
-  <div class="footer-copyright text-center py-3">
-    <p><a href="https://github.com/radualexandrub/SpringBoot-Angular-EmployeeManagerApp#running-the-app-locally"
-        target="_blank" title="See the code on GitHub"><i class="fa fa-github fa-2x"></i></a></p>
-    <p>Employee Manager App powered by <a href="https://angular.io/" target="_blank">Angular</a> &amp; <a
-        href="https://spring.io/projects/spring-boot" target="_blank">Spring Boot</a></p>
-    <p>Copyright 2022 © <a href="https://radubulai.com" target="_blank">Radu-Alexandru B</a></p>
-  </div>
-</footer>
+<app-footer></app-footer>
