+++ employeemanager-angular/src/app/components/footer/footer.component.html
@@ -1,0 +1,23 @@
+<footer class="page-footer font-small blue">
+  <div class="footer-copyright text-center py-3">
+    <p>
+      <a
+        href="https://github.com/radualexandrub/SpringBoot-Angular-EmployeeManagerApp#running-the-app-locally"
+        target="_blank"
+        title="See the code on GitHub"
+        ><i class="fa fa-github fa-2x"></i
+      ></a>
+    </p>
+    <p>
+      Employee Manager App powered by
+      <a href="https://angular.io/" target="_blank">Angular</a> &amp;
+      <a href="https://spring.io/projects/spring-boot" target="_blank"
+        >Spring Boot</a
+      >
+    </p>
+    <p>
+      Copyright 2022 &copy;
+      <a href="https://radubulai.com" target="_blank">Radu-Alexandru B</a>
+    </p>
+  </div>
+</footer>
