+++ employeemanager-angular/src/app/components/header/header.component.ts
@@ -1,6 +1,6 @@
-import { Component, OnInit } from '@angular/core';
-import { AppComponent } from 'src/app/app.component';
+import { Component, Injectable, OnInit } from '@angular/core';
 import { Employee } from 'src/app/employee';
+import { EmployeesComponent } from '../employees/employees.component';
 
 @Component({
   selector: 'app-header',
@@ -8,33 +8,33 @@ import { Employee } from 'src/app/employee';
   styleUrls: ['./header.component.css'],
 })
 export class HeaderComponent implements OnInit {
-  constructor(private appComponent: AppComponent) {}
+  constructor(private employeesComponent: EmployeesComponent) {}
 
   ngOnInit(): void {
-    if (localStorage.getItem("theme") === null) {
-      localStorage.setItem("theme", "light");
-      document.documentElement.setAttribute("data-theme", "light")
-    } else if (localStorage.getItem("theme") === "dark") {
-      document.documentElement.setAttribute("data-theme", "dark")
+    if (localStorage.getItem('theme') === null) {
+      localStorage.setItem('theme', 'light');
+      document.documentElement.setAttribute('data-theme', 'light');
+    } else if (localStorage.getItem('theme') === 'dark') {
+      document.documentElement.setAttribute('data-theme', 'dark');
     }
   }
 
   onToggleDarkTheme(): void {
-    const currentTheme = localStorage.getItem("theme");
-    if (currentTheme === "light") {
-      localStorage.setItem("theme", "dark");
-      document.documentElement.setAttribute("data-theme", "dark");
+    const currentTheme = localStorage.getItem('theme');
+    if (currentTheme === 'light') {
+      localStorage.setItem('theme', 'dark');
+      document.documentElement.setAttribute('data-theme', 'dark');
     } else {
-      localStorage.setItem("theme", "light");
-      document.documentElement.setAttribute("data-theme", "light");
+      localStorage.setItem('theme', 'light');
+      document.documentElement.setAttribute('data-theme', 'light');
     }
   }
 
   onOpenModal(employee: Employee, mode: string): void {
-    this.appComponent.onOpenModal(employee, mode);
+    this.employeesComponent.onOpenModal(employee, mode);
   }
 
   searchEmployees(keyword: string): void {
-    this.appComponent.searchEmployees(keyword);
+    this.employeesComponent.searchEmployees(keyword);
   }
 }
