+++ employeemanager-angular/src/app/components/modals/add-employee-modal/add-employee-modal.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { NgForm } from '@angular/forms';
-import { AppComponent } from 'src/app/app.component';
-import { Employee } from 'src/app/employee';
+import { EmployeesComponent } from '../../employees/employees.component';
 
 @Component({
   selector: 'app-add-employee-modal',
@@ -9,11 +8,11 @@ import { Employee } from 'src/app/employee';
   styleUrls: ['./add-employee-modal.component.css'],
 })
 export class AddEmployeeModalComponent implements OnInit {
-  constructor(private appComponent: AppComponent) {}
+  constructor(private employeesComponent: EmployeesComponent) {}
 
   ngOnInit(): void {}
 
   public onAddEmployee(addForm: NgForm): void {
-    this.appComponent.onAddEmployee(addForm);
+    this.employeesComponent.onAddEmployee(addForm);
   }
 }
